refactor(text): move highlighters into subdir of text
