+++ bin/CMakeLists.txt
@@ -35,7 +35,7 @@ make_target(
         DEPENDS
         TextFlow cli about macro pas builtins isa targets sim
         Qt6::Core Qt6::Gui Qt::Qml Qt::Quick Qt::QuickControls2
-        Qt::Widgets test-lib-all memory highlight
+        Qt::Widgets test-lib-all memory text
 )
 
 target_compile_definitions(pepp PRIVATE "INCLUDE_GUI=1")
