+++ bin/commands/gui.cpp
@@ -28,6 +28,7 @@
 #include "../gui/helpview/registration.hpp"
 #include "../gui/textedit/registration.hpp"
 #include "memory/hexdump/memorybytemodel.h"
+#include "text/plugin.hpp"
 struct default_data : public gui_globals {
   ~default_data() override = default;
   StatusBitModel sbm;
@@ -37,6 +38,7 @@ struct default_data : public gui_globals {
 };
 
 QSharedPointer<gui_globals> default_init(QQmlApplicationEngine &engine) {
+  text::registerTypes("edu.pepp");
   textedit::registerTypes(engine);
   helpview::registerTypes(engine);
   //  Note, these models are instantiated in C++ and passed to QML. QML
