+++ ui/CMakeLists.txt
@@ -1,11 +1,11 @@
-add_subdirectory(highlight)
+add_subdirectory(text)
 add_subdirectory(memory)
 
-set(ui_targets "memory;highlight")
+set(ui_targets "memory;text")
 
 
-if(APPLE)
+if (APPLE)
     install(TARGETS ${ui_targets} LIBRARY DESTINATION pepp.app/Contents/Frameworks)
-else()
+else ()
     install(TARGETS ${ui_targets} RUNTIME DESTINATION bin)
-endif()
+endif ()
