+++ ui/text/highlight/qml_highlighter.cpp
