+++ ui/text/highlight/style/defaults.cpp
