+++ ui/text/highlight/syntax/patternedhiglighter.cpp
