+++ ui/text/highlight/syntax/patternedhiglighter.hpp
