+++ ui/text/highlight/syntax/rules_clike.cpp
